src/main/java/org/whyspring/beans/factory/annotation/InjectionMetadata.java
package org.whyspring.beans.factory.annotation;

import java.util.List;

public class InjectionMetadata {

    // 目标类（为它的属性注入值）
    private final Class<?> targetClass;

    // 需要注入值的属性集合
    private List<InjectionElement> injectionElements;

    public InjectionMetadata(Class<?> targetClass, List<InjectionElement> injectionElements) {
        this.targetClass = targetClass;
        this.injectionElements = injectionElements;
    }

    public List<InjectionElement> getInjectionElements() {
        return injectionElements;
    }

    public void inject(Object target) {
        if (injectionElements == null || injectionElements.isEmpty()) {
            return;
        }
        for (InjectionElement ele : injectionElements) {
            ele.inject(target);
        }
    }
}
